modification

+++ src/hjem/is/db/StoragePlanSqlStore.java
@@ -146,6 +146,9 @@ public class StoragePlanSqlStore implements IStoragePlanStore {
 
     @Override
     public void delete(StoragePlan storagePlan) throws DataAccessException {
+        if (storagePlan.getId() == null) {
+            throw new IllegalArgumentException("this object isnt in the database and cant be updated");
+        }
         try {
             PreparedStatement stmt = DBConnection.getInstance().getConnection().prepareStatement("DELETE FROM storage_plans WHERE id = ?");
             stmt.setInt(1, storagePlan.getId());
